modifiy of values and data type of percentiles into summary process

+++ fenix-D3S-gift/src/main/java/org/fao/fenix/d3p/process/impl/dto/Percentile.java
@@ -2,12 +2,12 @@ package org.fao.fenix.d3p.process.impl.dto;
 
 
 public class Percentile {
-    public int percentile;
+    public double percentile;
     public double value;
 
     public Percentile() {
     }
-    public Percentile(int percentile, double value) {
+    public Percentile(double percentile, double value) {
         this.percentile = percentile;
         this.value = value;
     }
